+++ src/main/java/quikserve/challenge/promotions/model/Rule.java
@@ -2,7 +2,6 @@ package quikserve.challenge.promotions.model;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,5 +9,5 @@ public class Rule {
 
     private List<String> products;
 
-    private BigDecimal minValue;
+    private Integer minValue;
 }
